fix windows build bug

+++ Rocket/Engine/Utils/HashFunction.h
@@ -21,9 +21,9 @@ namespace Rocket {
         constexpr uint64_t u64(const char *s, uint32_t v) { return *s ? u64(s + 1, v * 256 + *s) : v; }
     }
     
-    constexpr int32_t operator "" _i16(const char *s, size_t) { return details::i16(s, 0); }
+    constexpr int16_t operator "" _i16(const char *s, size_t) { return details::i16(s, 0); }
     constexpr int32_t operator "" _i32(const char *s, size_t) { return details::i32(s, 0); }
-    constexpr int32_t operator "" _i64(const char *s, size_t) { return details::i64(s, 0); }
+    constexpr int64_t operator "" _i64(const char *s, size_t) { return details::i64(s, 0); }
     constexpr uint16_t operator "" _u16(const char *s, size_t) { return details::u16(s, 0); }
     constexpr uint32_t operator "" _u32(const char *s, size_t) { return details::u32(s, 0); }
     constexpr uint64_t operator "" _u64(const char *s, size_t) { return details::u64(s, 0); }
